requirements.txt
selenium
pytest
pytest-html
pytest-rerunfailures
pyyaml